requirements.txt
fastapi==0.112.1
pydantic==2.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn==0.30.6